Initial Commit - Semantic Segmentation Helper Functions
Function to load the segmentation data (RGB weather images and corresponding pixel label images) in datastores. Function to partition the datastores into separate training, validation, and testing datastores before combining the images and labels (currently set to 60/20/20). Function to augment the training data with a random reflections and translations. reflectivityScale enumeration class governs the categories the labels can be set to. Updated encodeWeatherData to automatically generate the labels for each pixel in a weather image using the general reflectivity guidelines given by NOAA and the radar detected reflectivity values.